+++ Common/src/main/java/fuzs/fastitemframes/config/ClientConfig.java
@@ -11,6 +11,10 @@ public class ClientConfig implements ConfigCore {
     public boolean disableItemOffsetWhenInvisible = false;
     @Config(description = "Hides the text label shown for names items.")
     public boolean disableNameTagRendering = false;
-    @Config(description = "Increases the scale of rendered items in invisible item frames.")
-    public boolean largeItemsWhenInvisible = true;
+    @Config(description = {
+            "Pick the scale of rendered items in invisible item frames.",
+            "A scale value of 0.5 is equal to the default rendering of items in normal item frames."
+    })
+    @Config.DoubleRange(min = 0.05, max = 1.0)
+    public double itemScaleWhenInvisible = 0.75;
 }